+++ BankingAPI.Core/Entities/DebitCard.cs
@@ -6,7 +6,7 @@ public class DebitCard : BaseEntity
     public string CardNumber { get; init; }
     public string CardHolderName { get; init; }
     public DateTime ExpiryDate { get; init; }
-    public int Limit { get; init; }
+    public decimal Limit { get; init; }
     public short CVV { get; init; }
     public int AccountId { get; init; }
     public int CustomerId { get; init; }
